requirements.txt
python-telegram-bot>=20,<22
aiofiles>=24.0
python-dotenv>=1.0
